Improve quantity input handling and error notifications
- Change quantity input from parseFloat to parseInt for whole numbers only
- Adjust mobile error banner positioning (bottom-20 to bottom-24)
- Replace delete error alert() with proper error state and banner display
- Add dismissible error notification for portfolio deletion failures
- Improves UX with better error messaging and state management

+++ apps/web/src/app/[lang]/portfolio/new/components/AssetList.tsx
@@ -90,7 +90,7 @@ export function AssetList({ items, colors, onRemove, onUpdateQuantity }: AssetLi
                       type="number"
                       value={item.quantity}
                       onChange={(e) => {
-                        const value = parseFloat(e.target.value);
+                        const value = parseInt(e.target.value);
                         onUpdateQuantity(item.symbol, Math.max(0, value || 0));
                       }}
                       className="w-full text-center bg-transparent font-bold text-sm text-black dark:text-white focus:outline-none p-0"
